created a search component
